rag-bot/requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
google-generativeai